# docs/openapi.yaml
openapi: 3.0.2
info:
  title: Prime Studio API
  version: 0.1.0
paths:
...
  /path/{path_id}:
    get:
      summary: Get Path
      description: Get learning path from DynamoDB
      operationId: get_path_path__path_id__get
      parameters:
        - required: true
          schema:
            title: Path Id
          name: path_id
          in: path
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Path"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /path:
    put:
      summary: Put Path
      description: Put learning path and relations to DynamoDB
      operationId: put_path_path_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReqPathPutTransact"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    post:
      summary: Post Path
      description: Post learning path to DynamoDB
      operationId: post_path_path_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReqPathPost"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    delete:
      summary: Delete Path
      description: Delete learning path and relations from DynamoDB
      operationId: delete_path_path_delete
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReqPathDeleteTransact"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /tag:
    put:
      summary: Put Tag
      description: Put tag to DynamoDB
      operationId: put_tag_tag_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReqTagPut"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    post:
      summary: Post Tag
      description: Post tag to DynamoDB
      operationId: post_tag_tag_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReqTagPost"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    delete:
      summary: Delete Tag
      description: delete tag meta and video/tag relations
      operationId: delete_tag_tag_delete
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReqTagDelete"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /upload/status:
    get:
      summary: Get Upload Status
      description: Get upload status from DynamoDB
      operationId: get_upload_status_upload_status_get
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Get Upload Status Upload Status Get
                type: array
                items:
                  $ref: "#/components/schemas/UploadStatus"
    put:
      summary: Put Upload Status
      description: Put upload status to DynamoDB
      operationId: put_upload_status_upload_status_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReqUploadStatusPut"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResUploadStatus"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    post:
      summary: Post Upload Status
      description: Post upload status to DynamoDB
      operationId: post_upload_status_upload_status_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReqUploadStatusPost"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResUploadStatus"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
components:
  schemas:
...
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"
...
    Path:
      title: Path
      required:
        - id
        - indexKey
        - PK
        - SK
        - createdAt
        - createdUser
        - updatedAt
        - updatedUser
        - invalid
        - name
        - description
      type: object
      properties:
        id:
          title: Id
          type: integer
        indexKey:
          title: Indexkey
          type: string
        PK:
          title: Pk
          type: string
        SK:
          title: Sk
          type: string
        createdAt:
          title: Createdat
          type: string
        createdUser:
          title: Createduser
          type: string
        updatedAt:
          title: Updatedat
          type: string
        updatedUser:
          title: Updateduser
          type: string
        invalid:
          title: Invalid
          type: boolean
        note:
          title: Note
          type: string
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
        videos:
          title: Videos
          type: array
          items:
            $ref: "#/components/schemas/VideoOrder"
          default: []
      description: Learning Path response body
...
    ReqPathDeleteTransact:
      title: ReqPathDeleteTransact
      required:
        - PK
        - user
      type: object
      properties:
        PK:
          title: Pk
          type: string
        user:
          title: User
          type: string
    ReqPathPost:
      title: ReqPathPost
      required:
        - name
        - description
        - user
      type: object
      properties:
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
        note:
          title: Note
          type: string
          default: ""
        user:
          title: User
          type: string
      description: Post learning path request body
    ReqPathPutTransact:
      title: ReqPathPutTransact
      required:
        - PK
        - user
      type: object
      properties:
        PK:
          title: Pk
          type: string
        user:
          title: User
          type: string
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
        note:
          title: Note
          type: string
        invalid:
          title: Invalid
          type: boolean
        appended:
          title: Appended
          type: array
          items:
            type: string
          default: []
        removed:
          title: Removed
          type: array
          items:
            type: string
          default: []
        orders:
          title: Orders
          type: array
          items:
            $ref: "#/components/schemas/VideoOrder"
          default: []
      description: Put learningh path request body
    ReqTagDelete:
      title: ReqTagDelete
      required:
        - PK
        - user
      type: object
      properties:
        PK:
          title: Pk
          type: string
        user:
          title: User
          type: string
      description: Delete tag request body
    ReqTagPost:
      title: ReqTagPost
      required:
        - user
        - name
      type: object
      properties:
        user:
          title: User
          type: string
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
          default: ""
        note:
          title: Note
          type: string
          default: ""
      description: Post tag request body
    ReqTagPut:
      title: ReqTagPut
      required:
        - PK
        - user
      type: object
      properties:
        PK:
          title: Pk
          type: string
        user:
          title: User
          type: string
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
        note:
          title: Note
          type: string
        invalid:
          title: Invalid
          type: boolean
      description: Put tag request body
...
    ReqUploadStatusPost:
      title: ReqUploadStatusPost
      required:
        - uri
        - user
        - name
        - filename
        - status
      type: object
      properties:
        uri:
          title: Uri
          type: string
        user:
          title: User
          type: string
        name:
          title: Name
          type: string
        filename:
          title: Filename
          type: string
        status:
          title: Status
          type: string
      description: Post upload status request body
    ReqUploadStatusPut:
      title: ReqUploadStatusPut
      required:
        - uri
        - timestamp
        - status
      type: object
      properties:
        uri:
          title: Uri
          type: string
        timestamp:
          title: Timestamp
          type: string
        status:
          title: Status
          type: string
      description: Put upload status request body
...
    ResUploadStatus:
      title: ResUploadStatus
      required:
        - uri
        - timestamp
        - status
      type: object
      properties:
        uri:
          title: Uri
          type: string
        timestamp:
          title: Timestamp
          type: string
        status:
          title: Status
          type: string
      description: Response upload status
...
    UploadStatus:
      title: UploadStatus
      required:
        - PK
        - SK
        - id
        - indexKey
        - createdAt
        - createdUser
        - name
        - filename
        - status
      type: object
      properties:
        PK:
          title: Pk
          type: string
        SK:
          title: Sk
          type: string
        id:
          title: Id
          type: string
        indexKey:
          title: Indexkey
          type: string
        createdAt:
          title: Createdat
          type: string
        createdUser:
          title: Createduser
          type: string
        name:
          title: Name
          type: string
        filename:
          title: Filename
          type: string
        status:
          title: Status
          type: string
      description: Upload status
...
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            type: string
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
...
    VideoOrder:
      title: VideoOrder
      required:
        - uri
      type: object
      properties:
        uri:
          title: Uri
          type: string
        order:
          title: Order
          type: integer
          default: 0
      description: Video order contain learning path
...